% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_working_days.R
\name{calculate_working_days}
\alias{calculate_working_days}
\title{Calculate number of working days in the UK by month and year}
\usage{
calculate_working_days(start, end)
}
\arguments{
\item{start}{A date in the format "YYYY-MM-DD"}

\item{end}{A date in the format "YYYY-MM-DD"}
}
\value{
a data frame with the number of working days in each month and total for the year
}
\description{
Calculate number of working days in the UK by month and year
}
\examples{
calculate_working_days(start = "2024-01-01", end = "2024-12-31")
}
