% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_filepath_format.R
\name{correct_filepath_format}
\alias{correct_filepath_format}
\title{Correct File Path Formats To Read Into R}
\usage{
correct_filepath_format(filepath)
}
\arguments{
\item{filepath}{A raw string giving a file path with directories separated using backslashes. It must be given as a raw string, for example: r"(path_to_file)".}
}
\value{
A string giving a file path with backslashes converted to forward slashes
}
\description{
Correct File Path Formats To Read Into R
}
\examples{
path <- correct_filepath_format(r"(top_directory\folder\sub-folder\file.txt)")
}
