% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_filename.R
\name{create_filename}
\alias{create_filename}
\title{Create a new filename to avoid overwriting existing files}
\usage{
create_filename(export_location, filepath)
}
\arguments{
\item{export_location}{A filepath to the location the new file should be written to}

\item{filepath}{The name and extension of the new file}
}
\value{
A character vector giving a new filename with a counter appended to avoid overwriting any existing files
}
\description{
Create a new filename to avoid overwriting existing files
}
\examples{
create_filename(export_location = "C:/Documents", filepath = "new_doc.txt")
}
