% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_temporary_name.R
\name{generate_temporary_name}
\alias{generate_temporary_name}
\title{Generate a random name}
\usage{
generate_temporary_name()
}
\value{
A character vector starting "temp_" and followed by a random set of characters
}
\description{
Generate a random name
}
\examples{
new_name <- generate_temporary_name()
}
