% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_financial_year.R
\name{get_financial_year}
\alias{get_financial_year}
\title{Get financial year from date}
\usage{
get_financial_year(date)
}
\arguments{
\item{date}{A date, in the format "YYYY-MM-DD"}
}
\value{
Financial year as a character
}
\description{
Get financial year from date
}
\examples{
get_financial_year("2025-07-05")
}
