% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Negate \%in\% For Filtering}
\usage{
a \%notin\% b
}
\arguments{
\item{a}{a vector to be filtered}

\item{b}{a vector of values to use in the filter}
}
\value{
TRUE or FALSE
}
\description{
Negate \%in\% For Filtering
}
\examples{
df <- data.frame(col1 = c("x", "y", "z"))

dplyr::filter(df, col1 \%notin\% c("x", "y"))
}
