% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JDutils-package.R
\docType{package}
\name{JDutils-package}
\alias{JDutils}
\alias{JDutils-package}
\title{JDutils: Useful Functions For Small But Frequent Tasks}
\description{
Useful functions for a variety of random small, but frequently needed, tasks.
}
\author{
\strong{Maintainer}: Jenny Davies \email{jennyedavies@outlook.com}

}
\keyword{internal}
