% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_packages_used.R
\name{list_packages_used}
\alias{list_packages_used}
\title{List R Packages Used In A Project}
\usage{
list_packages_used()
}
\value{
A list of R packages used in the current working project in alphabetical order, printed to the console.
}
\description{
List R Packages Used In A Project
}
\examples{
list_packages_used()
}
