% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_union_files.R
\name{read_and_union_files}
\alias{read_and_union_files}
\title{Read and union multiple CSV files}
\usage{
read_and_union_files(filepath)
}
\arguments{
\item{filepath}{A relative or absolute filepath where the files to be combined are saved}
}
\value{
A dataframe containing the combined files
}
\description{
Read and union multiple CSV files
}
\examples{
df <- read_and_union_files(filepath = "C:/User/Downloads/Files_to_combine")
}
