% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_debugging_options.R
\name{set_debugging_options}
\alias{set_debugging_options}
\title{Set Debugging Options}
\usage{
set_debugging_options()
}
\value{
Nothing, but sets options
}
\description{
Set options to use the searcher package to automatically prompt ChatGPT with error messages using the "debugging" prompt.
}
\examples{
set_debugging_options()
}
