% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_quietly.R
\name{stop_quietly}
\alias{stop_quietly}
\title{Stop Quietly i.e. Without Error Messages}
\usage{
stop_quietly()
}
\value{
Nothing - used to stop a process without an error message, where stopping is the desired behaviour of the program
}
\description{
Stop Quietly i.e. Without Error Messages
}
\examples{
stop_quietly()
}
